+++ src/probe/mod.rs
@@ -1,11 +1,13 @@
 pub mod correlate;
 pub mod icmp;
+pub mod interface;
 pub mod socket;
 pub mod tcp;
 pub mod udp;
 
 pub use correlate::*;
 pub use icmp::*;
+pub use interface::*;
 pub use socket::*;
 pub use tcp::*;
 pub use udp::*;